Split Unit test file from main game feature file. Implemented autodoc and added docstrings
